refactor: Move types into their own folder for easier re-use
