+++ src/definers/blueprints.ts
@@ -1,29 +1,4 @@
-import type {BlueprintResource} from '../types.js'
-
-export interface BlueprintsApiConfig {
-  organizationId: string
-  projectId: string
-  stackId: string
-}
-
-export interface BlueprintOutput {
-  name: string
-  value: string
-}
-
-export interface Blueprint {
-  $schema: string
-  blueprintVersion: string
-  resources?: BlueprintResource[]
-  values?: Record<string, unknown>
-  outputs?: BlueprintOutput[]
-}
-
-export type BlueprintModule = ((args?: unknown) => Blueprint) & {
-  organizationId?: string
-  projectId?: string
-  stackId?: string
-}
+import type {Blueprint, BlueprintModule, BlueprintsApiConfig} from '../index.js'
 
 export function defineBlueprint(blueprintConfig: Partial<Blueprint> & Partial<BlueprintsApiConfig>): BlueprintModule {
   const {organizationId, projectId, stackId, blueprintVersion, resources, values, outputs} = blueprintConfig
