+++ src/definers/functions.ts
@@ -1,33 +1,11 @@
-import type {BlueprintResource} from '../types'
-
-// --- Function Event Types ---
-
-export interface BlueprintFunctionBaseResourceEvent {
-  on?: [BlueprintFunctionResourceEventName, ...BlueprintFunctionResourceEventName[]]
-  filter?: string
-  projection?: `{${string}}`
-  includeDrafts?: boolean
-}
-export interface BlueprintDocumentFunctionResourceEvent extends BlueprintFunctionBaseResourceEvent {
-  includeAllVersions?: boolean
-  resource?: BlueprintFunctionResourceEventResourceDataset
-}
-export interface BlueprintMediaLibraryFunctionResourceEvent extends BlueprintFunctionBaseResourceEvent {
-  resource: BlueprintFunctionResourceEventResourceMediaLibrary
-}
-export type BlueprintFunctionResourceEvent = BlueprintDocumentFunctionResourceEvent | BlueprintMediaLibraryFunctionResourceEvent
-
-interface BlueprintFunctionResourceEventResourceDataset {
-  type: 'dataset'
-  /** @description A dataset ID in the format <projectId>.<datasetName>. <datasetName> can be `*` to signify "all datasets in project with ID <projectId>." */
-  id: string
-}
-interface BlueprintFunctionResourceEventResourceMediaLibrary {
-  type: 'media-library'
-  id: string
-}
-
-type BlueprintFunctionResourceEventName = 'publish' | 'create' | 'delete' | 'update'
+import type {
+  BlueprintBaseFunctionResource,
+  BlueprintDocumentFunctionResource,
+  BlueprintDocumentFunctionResourceEvent,
+  BlueprintFunctionBaseResourceEvent,
+  BlueprintMediaLibraryAssetFunctionResource,
+  BlueprintMediaLibraryFunctionResourceEvent,
+} from '../index.js'
 
 type BaseFunctionEventKey = keyof BlueprintFunctionBaseResourceEvent
 const BASE_EVENT_KEYS = new Set<BaseFunctionEventKey>(['on', 'filter', 'projection', 'includeDrafts'])
@@ -36,25 +14,8 @@ const DOCUMENT_EVENT_KEYS = new Set<DocumentFunctionEventKey>(['includeAllVersio
 type MediaLibraryFunctionEventKey = keyof BlueprintMediaLibraryFunctionResourceEvent
 const MEDIA_LIBRARY_EVENT_KEYS = new Set<MediaLibraryFunctionEventKey>(['resource', ...BASE_EVENT_KEYS.values()])
 
-// --- Main Function Types ---
-
 interface RequiredFunctionProperties {
   name: string
-}
-export interface BlueprintBaseFunctionResource extends BlueprintResource {
-  displayName?: string
-  src: string
-  timeout?: number
-  memory?: number
-  env?: Record<string, string>
-}
-export interface BlueprintDocumentFunctionResource extends BlueprintBaseFunctionResource {
-  type: 'sanity.function.document'
-  event: BlueprintDocumentFunctionResourceEvent
-}
-export interface BlueprintMediaLibraryAssetFunctionResource extends BlueprintBaseFunctionResource {
-  type: 'sanity.function.media-library.asset'
-  event: BlueprintMediaLibraryFunctionResourceEvent
 }
 
 export function defineDocumentFunction(
