+++ src/definers/roles.ts
@@ -1,28 +1,4 @@
-import type {BlueprintResource} from '../types'
-
-export interface RolePermission {
-  name: string // Required: predefined permission name (e.g., 'sanity-all-documents')
-  action: string // Required: permission action (e.g., 'read', 'mode')
-  params?: Record<string, unknown> // Optional: additional parameters for the permission
-}
-
-export interface BlueprintRoleConfig {
-  name: string
-  title: string
-  description?: string
-  appliesToUsers: boolean
-  appliesToRobots: boolean
-  permissions: RolePermission[]
-}
-
-export interface BlueprintRoleResource extends BlueprintRoleConfig, BlueprintResource {
-  type: 'sanity.access.role'
-}
-
-export interface BlueprintProjectRoleResource extends BlueprintRoleResource {
-  resourceType: 'project'
-  resourceId: string
-}
+import type {BlueprintProjectRoleResource, BlueprintRoleConfig, BlueprintRoleResource} from '../index.js'
 
 /**
  * Defines a role that is scoped to the same resource as the blueprint.
