+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './types.js'
+// DEFINERS - helper functions to create resources
 export * from './definers/blueprints.js'
 export * from './definers/cors.js'
 export * from './definers/datasets.js'
@@ -6,3 +6,12 @@ export * from './definers/functions.js'
 export * from './definers/resources.js'
 export * from './definers/roles.js'
 export * from './definers/webhooks.js'
+
+// TYPES - base types for all resources
+export * from './types/blueprints.js'
+export * from './types/cors.js'
+export * from './types/datasets.js'
+export * from './types/functions.js'
+export * from './types/resources.js'
+export * from './types/roles.js'
+export * from './types/webhooks.js'
